improuvement

+++ Multi-scale aperture robust optical flow/main.py
@@ -11,9 +11,9 @@ path_file_data = os.path.join(current_path,'data', name_data_file)
 data = loadmat(path_file_data)
 
 # Access to the data in the .mat file
-ts = data['ts']
-x = data['x']
-y = data['y']
+ts = data['ts'].astype(np.uint8)
+x = data['x'].astype(float)
+y = data['y'].astype(float)
 
 
 flow_local, corrected_flow = multi_scale_aperture_robust_optical_flow(x, y, ts)
